Move unit tests to tests/unit
